Add Python set of scripts to send bulk emails
The sender SMTP server and login credentials are configured in the
email.conf file.

Destination emails and configuration keys to be replaced in message
template are located in a CSV file. The CSV file is read, email messages
are constructed, and emails are sent.

Configuration allows for a pause between consecutive messages to prevent
marking the SMTP server as spam sender.

Signed-off-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>